Add support for jumps
Added multiplexer to choose the next value of the program
counter. One input is PC+2, the other the jump target.
No support to make conditional jumps depending on ALU results
yet.
Add a jump to the test program. The jump executes correctly,
although the instruction after the jump is also executed before
the jump is actually done.