Block loggers
Bind a badly connected logger to the exec unit from the exec unit
testbench.
On reset the program counter is now set to 0 via a mux
Remove support for MOVRR, as it is not needed.
Change JZI for JNZI. Makes more sense to check for non-zero.
Played a bit with the implementation settings. No idea.