Improved verification
The register file scoreboard no longer listens for register file writes.
As it listens to which instructions have been injected, it now keeps
track of the register values based on that, and compares the register
file contents at the end of each executed instruction.
The memory driver now injects ADDRR instructions as well, and can
easily be extended to all the rest of the instruction set.