Add a memory bus checker
This checker snoops on the memory interface and decodes the instructions
in flight. The sets the next expected read address according to the
decoded instruction. If the instruction is not a JNZ, the next memory
access (barring LOADs, need tro add support for them) should be the
start address of the current instruction + 2. If it´s a jump, it checks
the value of the ZERO bit in the execution unit and decides whether the
jump will be taken or not. If taken, set the next expected read address
to the jump target.