Initial idea to improve LOAD/STORE
Having hardcoded 8-bit offsets in LOAD and STORE does not really help
when you want to iterate over a memory range. It would be much better
to have the r0 register for the most significant 8 bits and another one
for the least significant 8 bits. This frees 4 bits in the instruction
encoding that could be used as an offset to access diferent fields in a
struct with changes in the addressing registers.